imageConverter/requirements.txt
MarkupSafe==3.0.2
mypy==1.15.0
mypy-extensions==1.0.0
pillow==11.1.0
typing_extensions==4.12.2
Werkzeug==3.1.3
